--- a/dlg_make_keil5_env_v2.000.py
+++ b/dlg_make_keil5_env_v2.000.py
@@ -379,7 +379,7 @@ def build_uvprojx_element_file(xml_sub_element, xml_tag, working_dir):
                 else:
                     # print("WARNING :: IT IS AN INVALID DIRECTORY PATH, THIS PATH WILL BE AUTOMATICALLY REMOVED...")
                     pass
-                print(single_text)
+                #print(single_text)
     print("")
 
     # my_file = open(DLG_UVPROJX_NAME,"w")
@@ -586,7 +586,7 @@ def build_uvprojx_element_various_controls(xml_sub_element, xml_tag, working_dir
 
     # Print updated_data in a readable format.
     updated_data = updated_data.replace(";", ";\r\n")
-    print("IncludePath ELEMENT TEXT:\r\n" + updated_data + "END OF IncludePath ELEMENT TEXT.\r\n")
+    #print("IncludePath ELEMENT TEXT:\r\n" + updated_data + "END OF IncludePath ELEMENT TEXT.\r\n")
 
     # print(temp_text)
     # print(updated_data[:-1])
@@ -841,7 +841,7 @@ def verify_dlg_sdk_proj_env_directory(path):
     else:
         return True
 
-L
+
 def handle_space_in_path(path):
     """
     For every directory in given path with a space in the string, this function will replace the