--- a/dlg_make_keil5_env_v2.000.py
+++ b/dlg_make_keil5_env_v2.000.py
@@ -1010,7 +1010,6 @@ def run_application(sdkpath):
 
 # start application
 if __name__ == "__main__":
-    starttime=time.time()
     # construct the argument parse and parse the arguments
     ap = argparse.ArgumentParser()
     ap.add_argument("-sdkpath", "--sdkpath", required=True,
@@ -1019,7 +1018,4 @@ if __name__ == "__main__":
 
     # print("SDK location {},is the sdk path".format(args["sdkpath"]))
     run_application(args["sdkpath"])
-    endtime=time.time()
-    difftime=endtime-starttime
-    print("script took "+str(difftime)+"s....")
 # run_individual_project_file(dirnames,args["sdkpath"])