--- a/dlg_make_keil5_env_v2.000.py
+++ b/dlg_make_keil5_env_v2.000.py
@@ -72,7 +72,6 @@ import re
 import argparse
 import xml.etree.ElementTree as ET
 from typing import NamedTuple
-import time
 
 '''
 Classes