The performance of the link script was pretty bad due to a lot of print statements. This PR removes those and the result is:
- cleaning took 107 seconds and now takes 43s
- linking took 49s and now takes 23s
The exact numbers depend on multiple factors such as the system it runs on and the lenght of the paths to print but this is roughly a 50% performance increase.